snap: add ref count mechanism for RDMA devices for sdev
This is needed only for BF-1 because of the special steering rules that
are used to implement device emulation. Also set the address of the
emulated function to be the same as the networking function (as soon as
possible) for loopback connections (use gid idx 0). For BF-2 and above,
many networking functions can be used for moving packes to/from the
emulated function due to the support of the HW for cross-gvmi access
(cross-gvmi mkey feature).

Signed-off-by: Max Gurtovoy <maxg@mellanox.com>