snap: add support for creating HW qp
HW qp is a qp that created on top of the emulated function namespace.
This is needed for Bluefield-1 based emulation. In order to send packets
with data from host memory, we need to create a qp that is associated to
a PCI function that resides in the host. For that we create a dedicated
qp and we clone a SW qp that is connected to the target qp. During this
operation, the user must make sure that there will not be any traffic.
E.g in the NVMe/NVMf emulation case, this operation can happen after
connection established in the NVMf layer (after we get successful
connect response on the specific queue). For Bluefiled-2 based
emulation, this won't be needed since we'll use cross-function (gvmi)
memory key to access host memory.

Signed-off-by: Max Gurtovoy <maxg@mellanox.com>