snap: add steering rules for Bluefield-1 based emulation
For Bluefield-1 SNAP, we must set special steering rules in the HW
e-switch in order to wire packets to/from QPs that reside in the NVMe PF
namespace. For Bluefield-2 and above, the QPs will be part of the RDMA
interface namespace within the SNIC and it will use cross-gvmi mkey in
order to access host memory.

Signed-off-by: Max Gurtovoy <maxg@mellanox.com>