snap: remove the need for custom rdma-core
Use mlx5dv_open_device instead of ibv_open_device. In case we open the
ibdev inside the library, we can use the dv API for that. In the future,
rdma-core will enable devx on all supported devices. This will enable
users to pass ibv_context and/or ibv_qp that was created by common verbs
API and the library will be able to manipulate them with direct verbs
API. Till than, user must make sure that all resources that created
outside the library will be created with devx enabled (instead of using
a custom rdma-core with always_devx enabled).

Signed-off-by: Max Gurtovoy <maxg@mellanox.com>