virtio: Detect virtio reset based on enabled bit
According to new PRM, changes in `device_status` should
not be considered as triggers for virtio reset, due to
possible race conditions with virtio v0.95 drivers.
Instead, always count on the internal `enabled` bit -
when set to `0`, reset (or FLR) happened, and controller
should set it back to `1` once finished resetting itself.

Signed-off-by: Nitzan Carmi <nitzanc@mellanox.com>