blk: Change API to single context with cb and user arg
The translation from snap bdev operation to bdev implementation might
need to perform extra steps after the bdev op finished (e.g. release io
memory). This means the translation can not provide the snap bdev cb as
the cb to the operation implementation, instead it needs to give it as a
user argument and call it later on (in its own cb). By getting both snap
cb and user argument as a single context the translation can pass it as a
single user argument to the implementation without the need to allocate
new memory to hold both and free it later.

Signed-off-by: Guy Heller <heller@mellanox.com>