snap: Include entire BDF in pci_number attribute
When moving to work with emulated switch, different
emulations might be opened on different PCI device/bus,
so move pci_number to include all PCI BDF (Bus, Device,
Function) instead of only the PCI function.
Luckily, the whole BDF fits into 16bit value, and function
number is part of BDF, so it is 100% compatible with BF-1.

Signed-off-by: Nitzan Carmi <nitzanc@mellanox.com>