rpc: replace pci_id with pci_bdf
In order to have a clear understanding of the mapping between the
emulated function instance in the Bluefield to the function shown on the
host, we've added pci_bdf representation to the PRM. Update the RPC API
to work with BDFs instead of pci index. Also, while we're here, add a
new RPC command to hotplug NVMe emulated functions.

Signed-off-by: Max Gurtovoy <maxg@mellanox.com>