snap: Workaround to avoid race in querying PCI BDF
Currently there is race condition where right after the hot plug, device
discovery on the host is not yet completed. So PCI BDF are not yet
available to the device.
Due to this, when query function info is done, PCI BDF are returned
zero.

Due to this user is in blind zone where link between PCI device and
its representor is missing.

issue: 2268268
Signed-off-by: Parav Pandit <parav@nvidia.com>