RPC: make params optional on hotplug/ctrl creation
When hotplug a device, we might want to add it with
default parameters, which are not related to specific bdev.
It is up to the virtio-blk controller to set the right parameters
before finalizing the "RESET" flow.
There is also a possibility a controller will be opened with
no real backend bdev (e.g. 0-size capacity), like in PF opened
only to be used for VFs management.

Signed-off-by: Nitzan Carmi <nitzanc@nvidia.com>