snap_nvme: Separate cross-gvmi mkey creation
Cross-gvmi mkey allows having an mkey that points to another mkey that
resides in a different vhca_id. This allows one vhca, using a locally
defined mkey, to access the memory of another vhca. It is very useful
when the other VHCA happens to be on a different host, as is the case in
emulated functions.

Up until now both the pointing key and the pointed key were created
together by FW when creating an NVME_SQ. By creating the keys separately
we can first create the pointed key, and only when we know on which
rdma-device we go out from we can create the pointing key.

Signed-off-by: Guy Heller <heller@mellanox.com>
Signed-off-by: Israel Rukshin <israelr@nvidia.com>