snap_channel: add rdma_cm create/destroy logic
Initially we'll use rdma_cm QP as the transport to pass commands
between the host and the SNAP controller on the Bluefield. For now, add
new env variables to set the ip and port that should be used for
creating the listener object.

Also, while we're here, define the new channel command interface that
should be used vs. the host. This can be used for creating new unit
tests for the rdma_cm part of the channel.

Signed-off-by: Max Gurtovoy <mgurtovoy@nvidia.com>