virtio: Fix hang at hotplug when modules are up
When hotplugging virtio-blk PF when host modules
are already up, there is a race in which sometimes
host PCI enumeration is blocked until virtio-blk
controller is created.
To avoid this deadlock, which only the user can solve,
set a short but reasonable timeout for PCI enumeration
waiting time, and print a detalied explanation about
needed steps to avoid this problem.

Signed-off-by: Nitzan Carmi <nitzanc@nvidia.com>