virtio: introduce progress lock
The commit replaces virtio controller state lock with a
progress lock. The problem with the state lock is that
it is too fine grained.

The lock should suspend the whole progress() and not just
snap_virtio_ctrl_stop/start

This is done in preparation for the live migration. The migration
channel may execute in its own thread context so it is necessary
to stop progress while executing suspend/resume and state
save/restore.

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>