virtio-blk: use inline rdma write to write status
Inline rdma write is better because it eliminates pci
access to get data. Also the overall request processing
is done a little faster because completion is now sent
immediately after the status is written.

It works because operations on the same QP are done in order.

Performance gain as tested with the bdevperf and local null bdev:

    gga
nqueues baseline    write_status_inl    %gain
1       408         423                 0.036764706
2       789         820                 0.039290241
4       1269        1473                0.160756501
8       1849        2133                0.153596539
16      1738        1953                0.123705409

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>