nvme: Avoid using snap_device during ctrl counters creation
We currently use snap_device during ctrl counters creation
and deletion, which causes us to destroy counters before
we release the device. This raises FW assert as we destroy
object after its counters are destroyed.

Lucky for us, we can avoid using the sdev, and that way
overcome that anomaly.

Signed-off-by: Nitzan Carmi <nitzanc@nvidia.com>