Virtio: Core allocation optimization
Same solution as in nvmx for nvme commit dba69f2fd.
This is a preliminary solution.
It doesn't take into account actual core load.
Instead, it just balances amount of queues per core.
The balancing is performed during VQ creation only.
It doesn't re-balance on the fly when queues are removed.

Signed-off-by: Dana Benbasat <dbenbasat@nvidia.com>