snap_dma: provide udp_sport when change QP to RTR state
Need to set rmac and udp_sport filed to Queue Pair Context
obj when set QP state to RTR in non-force_loopback mode.
And, we are work with RoCE v2, it require a non-zero
udp_sport value, so need to calculate a udp_sport for it.
For RC type RDMA connection, use local QP number and remote
QP number to calc a udp_sport.

Signed-off-by: Tom Wu <tomwu@nvidia.com>