dpa: add apu specific makefile and example code
The commit adds makefile that supports cross compilation
for the DPA. Code from the Flexio SDK is used as a proof of
concept DPA application. The code is picked as is and not
formatted to the snap style.

Dummy dpa_hello.c is used to test our cross compilation
environment.

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>