Split `virtq_read_req_from_host` function
The `virtq_read_req_from_host` function does several things
 - Compress descriptors
 - Allocates extra I/O buffer if required
 - Load header
 - Load request data (if any)

When use a memory pool, we want to minimize the time we own the
buffer. So we can benefit from splitting the function and add
more states to the state machine.

We'll be able allocate the memory right before we need it.
We also will skip allocations for commands which don't need
I/O buffers.

Also, the code becomes more self-documented since functions
are doing exactly what their names mean.

Signed-off-by: Eugene Astafiev <eastafiev@nvidia.com>