virtio-blk: fix for bug #2747508
Started from release 08/2021, the recovery flag
should be used as default during the creation of
the controller. We need to distinguish between 'real'
recovery or 'new creation' of the controller.
This can be done by testing the reset flag.
For 'new creation' it should be as following:
enabled=1 reset=1 status=0

When the reset flag is turned on a new controller should
be created, otherwise - proceed with the recovery.

Tests:
- Create null bdev
- Create new hot-plug emulation
- Create controller attached into bdev & emulation above
- Load virtio-blk module on host
- Run lsblk - you should see a new block device 'vdxx'

Signed-off-by: Rostislav Raikhman <rraikhman@nvidia.com>