virtio_blk: Fix crash when zcopy=1 and mempool=1
If both zcopy and mempool are configured, we don't allocate buffers
from memory pool because I/O operations don't use them due to
zcopy nature.

However, the GET_ID command stays aside and still needs intermediate
buffer for requesting the ID from bdev. It always behave the same
regardless of zcopy and mempool options.

In order to provide a buffer to the GET_ID command, the zcopy_check
function is extended with additional check. It disables zcopy for
the GET_ID cmd. The flow then fallbacks to either mempool or
static allocation model which are both ok.

Signed-off-by: Eugene Astafiev <eastafiev@nvidia.com>