virtio_blk: change virtq_ctx_init prototype
This commit changes the prototype of virtq_ctx_init
function:
Added the tx_elem_size for the virtq_ctx_init
function. The parameter, later, will be used as a hint
during QP creation for desired 'inline RDMA send length.
Till now, for blk, it was the size of the tunneling struct.
Note: the inline RDMA send is used for sending the blk
write status (byte) & tunneling data.

For virtio_fs the status is 16 bytes which
is >> size of tunneling struct.

As the number of parameters is pretty big fold them into
a helper struct: virtq_ctx_init_attr.

Signed-off-by: Rostislav Raikhman <rraikhman@nvidia.com>