virtio_blk: fix fatal_error processing
The fix is related to 'fatal_error' processing.
The fatal_error can occur in different scenarios like:
  a. invalid input data (data of request or virtio-chain)
  b. RDMA error during
        reading data from host (read descriptors or data)
        writing virito status
        writing tunneling completion
  c. allocation memory error during handling
        extra descriptors

In case of fatal error, viritio queue cannot complete the request
and as result, we have "outstanding uncomplete I/O".
The number of uncomplete I/Os might increase in case the controller
was created with the force_in_order flag. As the other completions
depends on I/Os which has a fatal error and cannot be completed.

In such a case, the user should be able to delete the controller.
The controller might be created later with recovery mode.

The issue: if virtio queue detects outstanding I/O
it waits for all completions first (which will never be happened)
and after that complete deletion of the queue.

Note: we have only one counter related to outstanding I/Os.

Fix:
separate the outstanding I/O counter to the following counters:
'incoming from host'
'sent to bdev'
'sent to host - counts read data/fetch descriptors/write data etc.'
'fatal errors'

When we want to delete the controller:
'sent to bdev' & 'sent to host' have to be 0
'incoming from host' value depends on 'fatal errors'.

Tests:
- configure controller with/without zcopy (with/without force_in_order)
- generate I/O
- simulate bad data (cause to fatal error)
- verify that snap is able to delete the controller
- create a controller with recovery mode
- verify that I/O continue without a problem

Signed-off-by: Rostislav Raikhman <rraikhman@nvidia.com>