snap: Avoid dependency of hotplug helpers on snap_context
To detach devx and low level commands from high level snap objects,
build helper APIs to accept only necessary arguments.

vhca_id in the helper routines should be uint16_t, but due to callers
using it for invalid vhca_id check as well, it is kept as int for now.

Signed-off-by: Parav Pandit <parav@nvidia.com>