virtio: Introduce snap_virtio_ctrl_is_configurable() API
We can only modify virtio PCI BAR as long as the driver
has not read it yet. We can only ensure such scenario
when DRIVER/FEATURES_OK/DRIVER_OK are not set.

Note that such modification is risky, and this is only the
first line of defence. There is still a small chance the
driver will read PCI BAR while we are modifying it. In that
rare case, we will be blocked by FW when trying to modify the
BAR with the new values later in the process.

Signed-off-by: Nitzan Carmi <nitzanc@nvidia.com>