dpa: workaround simx thread destruction bug
It looks like there is a deadlock in the riscv kernel when it tries to
destroy a thread that is still running.

This is what is seen in the log:
mlx: ERR-APU [mlx5/mlx5_cmdif_apu.c:92] mlx5/mlx5_cmdif_general_obj.c:893
_mlx5_apu_comm_channel_fw2apu_owner_wait(): Waiting more than 42040
seconds for APU to release ownership

I suspect that riscv kernel is stuck in thread_destroy. Most probably in
the cleanup_running_harts() busy loop.

Acked by TalG and ElijahS

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>