code_style: Ditch "Prefer ARRAY_SIZE" style check
checkpatch.pl was originally designed to be used over
linux kernel, where a built-in ARRAY_SIZE macro exists.
In userspace, there is a false warning about not using
the built-in ARRAY_SIZE(n) when in fact we just define the
userspace equivalent macro.

As we still need this check to be enabled for other places
in the code, we can easily ditch this warning by replacing
n[0] with *n.

Signed-off-by: Nitzan Carmi <nitzanc@nvidia.com>