virtio_blk: Change bdev detach to single command
Allow bdev to be detached (or deleted) without suspending ctrl or
unprobing driver.
Capacity can be changed while host driver is up.
When detaching bdev if the controller was not suspended wait for all
outstanding io commands to complete. All incoming commands are failed.
When there are no more outstanding io commands on the ctrl we can safely
detach the bdev.

Signed-off-by: Dana Benbasat <dbenbasat@nvidia.com>