snap: add support to read max num of VFs
For each emulated physical function, a new PRM filed is defined for
max num of VFs supported by this device. To maintain backward
compatibility, the valid bit must be true in order to make use of
the max number.

This number is useful for controller to reserve or pre-create resources
for VFs.

Signed-off-by: Bodong Wang <bodong@nvidia.com>