snap_dpa+dpa: add command cq
The commit adds a cq that receives wakeup events to
the DPA thread.

This is needed to support efficient slow path command handling in
the 'event' mode. It is going to be much faster to check
command cq for new cqes on each wakeup than to reconfigure window
and access mailbox memory.

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>