snap_dma: post recv after qp moved to the INIT state
According to the ib spec recv must be posted after QP
is moved to the INIT state. While our hardware does not
enforce this, new rdma-core (ofed5.6) has added this check.

The commit moves post recv until after the qps has been
connected (in RTS state).

This is OK becase both sides of the connection are local and
we know for sure that no packets are going to be sent until
snap_dma_ep_connect() or snap_dma_q_create() is completed.

Signed-off-by: Alex Mikheev <amikheev@nvidia.com>