snap: Enable capabilities for virtio net queue counters
Previously, this was only being done for virtio_blk devices. Extend this
functionality for virtio_net devices as well. Failure to do so would
cause an incorrect counter object id being assigned to a HW queue at the
time of queue creation. Eventually, when the queue counters are queried,
the statistics reported would be incorrect.

Issue: 3074110
Signed-off-by: Vivek Bhat <vivekb@nvidia.com>