virtio: fix in snap_vq_suspend function
As suspend is called in context of one thread & progress in context
of other thread we can meet the case when queue is moved to
suspended state but we still call progess_rx in context of the thread.

So, we will continue with suspend processing
in parallel of handling inflight I/Os.

Signed-off-by: Rostislav Raikhman <rraikhman@nvidia.com>