snap_dma/dv: enable TX Scatter-to-CQE by default
The enabled TX Scatter-to-CQE mode is up-to-32B.

This commit also fix snap_dma_q_read() API
when read data is less or equal to 32 bytes,
becasue in this case the respond data will
scatter to CQE if TX Scatter-to-CQE enabled,
but snap_dma_q_read() API expect the respond
data will put to the provided rx buffer. Do
a memcpy inside dma complation to fix this.

Signed-off-by: Tom Wu <tomwu@nvidia.com>