Vq reset (#825)
* snap/net: Support virtio queue config version 2

According to PRM for virtio queue config version 2, following are
added:

	1. new entries for queue configs in modify_field_select
	2. q_cfg_version to let FW knows if SW support queue
	   config version 2
	3. q_configuration_list_size to specify the number of
	   queues
	4. Virtio_Q_Configuration_V2_List

This patch added the support for virtio queue config version 2.

Signed-off-by: Wei Zhang <weizhang@nvidia.com>

* snap: Support virtio queue reset feature

To support virtio queue reset, a reset_check is added to progress
so that the reset flag can be checked. If reset flag for a
queue is true, the corresponding reset flag in queue attr is
marked true, and then it triggers a callback function call to
the controller to reset the queue. And after all queues in the
list are reset, a device_modify with queue_enable=0 is called to
notify FW that the reset has been finished.

Signed-off-by: Wei Zhang <weizhang@nvidia.com>

* snap: Support virtio on-demand queue enable feature

To support virtio queue enable, a enable_check is added to progress
so that the enable flag can be checked. If enable flag for a
queue is true, the corresponding enable flag in queue attr is
marked true, and then it triggers a callback function invoke to
the controller to enable (i.e. create) the queue.

Signed-off-by: Wei Zhang <weizhang@nvidia.com>

Co-authored-by: Wei Zhang <weizhang@nvidia.com>