snap: Fix Admin VQ creation for NET
Currently snap_vq_hwq_create does not take in account the fact that a
later call to snap_virtio_create_queue which will cast the vattr struct
to its net portion and fetch the HW available and used index.
To allow it, create a net spesific struct and handle this case properly.

Signed-off-by: Armen Ratner <armeng@nvidia.com>